@@ -9,29 +9,33 @@
 		"lint": "prettier --check . && eslint .",
 		"format": "prettier --write ."
 	},
-	"devDependencies": {
-		"@iconify/svelte": "^4.2.0",
-		"@sveltejs/adapter-auto": "^3.3.1",
-		"@sveltejs/adapter-static": "^3.0.8",
-		"@sveltejs/kit": "^2.20.2",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"@types/eslint": "^8.56.12",
-		"autoprefixer": "^10.4.21",
-		"eslint": "^8.57.1",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.46.1",
-		"flowbite": "^2.5.2",
-		"flowbite-svelte": "^0.46.23",
-		"postcss": "^8.5.3",
-		"prettier": "^3.5.3",
-		"prettier-plugin-svelte": "^3.3.3",
-		"svelte": "^4.2.19",
-		"tailwindcss": "^3.4.17",
-		"vite": "^5.4.15"
-	},
 	"type": "module",
 	"dependencies": {
-		"@fontsource-variable/jetbrains-mono": "^5.2.5",
-		"flowbite-svelte-icons": "^1.6.2"
+		"@eslint/compat": "^1.3.0",
+		"@eslint/js": "^9.18.0",
+		"@fontsource-variable/jetbrains-mono": "^5.2.6",
+		"@iconify/svelte": "^4.2.0",
+		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/adapter-static": "^3.0.8",
+		"@sveltejs/enhanced-img": "^0.4.4",
+		"@sveltejs/kit": "^2.21.4",
+		"@sveltejs/vite-plugin-svelte": "^5.1.0",
+		"@tailwindcss/forms": "^0.5.10",
+		"@tailwindcss/typography": "^0.5.16",
+		"@tailwindcss/vite": "^4.1.10",
+		"autoprefixer": "^10.4.21",
+		"eslint": "^9.28.0",
+		"eslint-config-prettier": "^10.1.5",
+		"eslint-plugin-svelte": "^2.46.1",
+		"flowbite": "^3.1.2",
+		"flowbite-svelte": "^0.47.4",
+		"globals": "^15.15.0",
+		"postcss": "^8.5.5",
+		"prettier": "^3.5.3",
+		"prettier-plugin-svelte": "^3.4.0",
+		"prettier-plugin-tailwindcss": "^0.6.12",
+		"svelte": "^5.34.1",
+		"tailwindcss": "^3.4.17",
+		"vite": "^6.3.5"
 	}
 }
