@@ -1,19 +1,15 @@
 <script>
-	import {
-		Footer,
-		FooterLinkGroup,
-		FooterLink,
-		Img,
-		Heading,
-		FooterCopyright
-	} from 'flowbite-svelte';
+	import { Footer, FooterLinkGroup, FooterLink, FooterCopyright } from 'flowbite-svelte';
 	import '../app.css';
 	import ScreenNav from '../lib/components/ScreenNav.svelte';
 	import PhoneNav from '../lib/components/PhoneNav.svelte';
 	import { base } from '$app/paths';
 	import '@fontsource-variable/jetbrains-mono';
 
-	let innerWidth = 0;
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
+
+	let innerWidth = $state(0);
 </script>
 
 <svelte:window bind:innerWidth />
@@ -26,7 +22,7 @@
 
 <div class="mt-24">
 	<div class="bg-slate-100 py-8">
-		<div><slot /></div>
+		<div>{@render children?.()}</div>
 	</div>
 	<Footer class="bg-slate-900 text-slate-300 py-8 text-center">
 		<FooterLinkGroup ulClass="flex justify-evenly mt-3 text-sm list-none p-0">
